+++ integration-test/src/test/java/org/apache/iotdb/db/it/IoTDBLoadTsFileIT.java
@@ -25,13 +25,10 @@ import org.apache.iotdb.db.it.utils.TestUtils;
 import org.apache.iotdb.it.env.EnvFactory;
 import org.apache.iotdb.it.framework.IoTDBTestRunner;
 import org.apache.iotdb.it.utils.TsFileGenerator;
-import org.apache.iotdb.it.utils.TsFileTableGenerator;
 import org.apache.iotdb.itbase.category.ClusterIT;
 import org.apache.iotdb.itbase.category.LocalStandaloneIT;
-import org.apache.iotdb.itbase.env.BaseEnv;
 import org.apache.iotdb.jdbc.IoTDBSQLException;
 
-import org.apache.tsfile.enums.ColumnCategory;
 import org.apache.tsfile.enums.TSDataType;
 import org.apache.tsfile.file.metadata.enums.TSEncoding;
 import org.apache.tsfile.read.common.Path;
@@ -108,8 +105,8 @@ public class IoTDBLoadTsFileIT {
     try (final Connection connection = EnvFactory.getEnv().getConnection();
         final Statement statement = connection.createStatement()) {
 
-      statement.execute("CREATE DATABASE " + SchemaConfig.STORAGE_GROUP_0);
-      statement.execute("CREATE DATABASE " + SchemaConfig.STORAGE_GROUP_1);
+      statement.execute("CREATE DATABASE " + SchemaConfig.DATABASE_0);
+      statement.execute("CREATE DATABASE " + SchemaConfig.DATABASE_1);
 
       statement.execute(convert2SQL(SchemaConfig.DEVICE_0, SchemaConfig.MEASUREMENT_00));
       statement.execute(convert2SQL(SchemaConfig.DEVICE_0, SchemaConfig.MEASUREMENT_01));
@@ -164,8 +161,8 @@ public class IoTDBLoadTsFileIT {
     try (final Connection connection = EnvFactory.getEnv().getConnection();
         final Statement statement = connection.createStatement()) {
 
-      statement.execute(String.format("delete database %s", SchemaConfig.STORAGE_GROUP_0));
-      statement.execute(String.format("delete database %s", SchemaConfig.STORAGE_GROUP_1));
+      statement.execute(String.format("delete database %s", SchemaConfig.DATABASE_0));
+      statement.execute(String.format("delete database %s", SchemaConfig.DATABASE_1));
     } catch (final IoTDBSQLException ignored) {
     }
   }
@@ -899,139 +896,9 @@ public class IoTDBLoadTsFileIT {
     return pairs;
   }
 
-  @Test
-  public void testLoadWithEmptyDatabaseForTableModel() throws Exception {
-    final int lineCount = 10000;
-
-    final List<Pair<MeasurementSchema, MeasurementSchema>> measurementSchemas =
-        generateMeasurementSchemasForDataTypeConvertion();
-    final List<ColumnCategory> columnCategories =
-        generateTabletColumnCategory(0, measurementSchemas.size());
-
-    final File file = new File(tmpDir, "1-0-0-0.tsfile");
-
-    final List<IMeasurementSchema> schemaList =
-        measurementSchemas.stream().map(pair -> pair.right).collect(Collectors.toList());
-
-    try (final TsFileTableGenerator generator = new TsFileTableGenerator(file)) {
-      generator.registerTable(SchemaConfig.TABLE_0, schemaList, columnCategories);
-
-      generator.generateData(SchemaConfig.TABLE_0, lineCount, PARTITION_INTERVAL / 10_000);
-    }
-
-    // Prepare normal user
-    try (final Connection adminCon = EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        final Statement adminStmt = adminCon.createStatement()) {
-      adminStmt.execute("create user test 'password123456'");
-      adminStmt.execute(
-          String.format(
-              "grant create, insert on %s.%s to user test",
-              SchemaConfig.DATABASE_0, SchemaConfig.TABLE_0));
-
-      // auto-create table
-      adminStmt.execute(String.format("create database if not exists %s", SchemaConfig.DATABASE_0));
-    }
-
-    try (final Connection connection =
-            EnvFactory.getEnv().getConnection("test", "password123456", BaseEnv.TABLE_SQL_DIALECT);
-        final Statement statement = connection.createStatement()) {
-      statement.execute(String.format("use %s", SchemaConfig.DATABASE_0));
-      statement.execute(String.format("load '%s'", file.getAbsolutePath()));
-    }
-
-    try (final Connection adminCon = EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        final Statement adminStmt = adminCon.createStatement()) {
-      adminStmt.execute(String.format("use %s", SchemaConfig.DATABASE_0));
-      try (final ResultSet resultSet =
-          adminStmt.executeQuery(String.format("select count(*) from %s", SchemaConfig.TABLE_0))) {
-        if (resultSet.next()) {
-          Assert.assertEquals(lineCount, resultSet.getLong(1));
-        } else {
-          Assert.fail("This ResultSet is empty.");
-        }
-      }
-    }
-  }
-
-  @Test
-  @Ignore("Load with conversion is currently banned")
-  public void testLoadWithConvertOnTypeMismatchForTableModel() throws Exception {
-    final int lineCount = 10000;
-
-    List<Pair<MeasurementSchema, MeasurementSchema>> measurementSchemas =
-        generateMeasurementSchemasForDataTypeConvertion();
-    List<ColumnCategory> columnCategories =
-        generateTabletColumnCategory(0, measurementSchemas.size());
-
-    final File file = new File(tmpDir, "1-0-0-0.tsfile");
-
-    List<MeasurementSchema> schemaList1 =
-        measurementSchemas.stream().map(pair -> pair.left).collect(Collectors.toList());
-    List<IMeasurementSchema> schemaList2 =
-        measurementSchemas.stream().map(pair -> pair.right).collect(Collectors.toList());
-
-    try (final TsFileTableGenerator generator = new TsFileTableGenerator(file)) {
-      generator.registerTable(SchemaConfig.TABLE_0, schemaList2, columnCategories);
-
-      generator.generateData(SchemaConfig.TABLE_0, lineCount, PARTITION_INTERVAL / 10_000);
-    }
-
-    try (final Connection connection =
-            EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        final Statement statement = connection.createStatement()) {
-      statement.execute(String.format("create database if not exists %s", SchemaConfig.DATABASE_0));
-      statement.execute(String.format("use %s", SchemaConfig.DATABASE_0));
-      statement.execute(convert2TableSQL(SchemaConfig.TABLE_0, schemaList1, columnCategories));
-      statement.execute(
-          String.format(
-              "load '%s' with ('database'='%s')", file.getAbsolutePath(), SchemaConfig.DATABASE_0));
-      try (final ResultSet resultSet =
-          statement.executeQuery(String.format("select count(*) from %s", SchemaConfig.TABLE_0))) {
-        if (resultSet.next()) {
-          Assert.assertEquals(lineCount, resultSet.getLong(1));
-        } else {
-          Assert.fail("This ResultSet is empty.");
-        }
-      }
-    }
-  }
-
-  private List<ColumnCategory> generateTabletColumnCategory(int tagNum, int filedNum) {
-    List<ColumnCategory> columnTypes = new ArrayList<>(tagNum + filedNum);
-    for (int i = 0; i < tagNum; i++) {
-      columnTypes.add(ColumnCategory.TAG);
-    }
-    for (int i = 0; i < filedNum; i++) {
-      columnTypes.add(ColumnCategory.FIELD);
-    }
-    return columnTypes;
-  }
-
-  private String convert2TableSQL(
-      final String tableName,
-      final List<MeasurementSchema> schemaList,
-      final List<ColumnCategory> columnCategoryList) {
-    List<String> columns = new ArrayList<>();
-    for (int i = 0; i < schemaList.size(); i++) {
-      final MeasurementSchema measurement = schemaList.get(i);
-      columns.add(
-          String.format(
-              "%s %s %s",
-              measurement.getMeasurementName(),
-              measurement.getType(),
-              columnCategoryList.get(i).name()));
-    }
-    String tableCreation =
-        String.format("create table %s(%s)", tableName, String.join(", ", columns));
-    LOGGER.info("schema execute: {}", tableCreation);
-    return tableCreation;
-  }
-
   private static class SchemaConfig {
-    private static final String DATABASE_0 = "root";
-    private static final String TABLE_0 = "test";
-    private static final String STORAGE_GROUP_0 = "root.sg.test_0";
-    private static final String STORAGE_GROUP_1 = "root.sg.test_1";
+    private static final String DATABASE_0 = "root.sg.test_0";
+    private static final String DATABASE_1 = "root.sg.test_1";
 
     // device 0, nonaligned, sg 0
     private static final String DEVICE_0 = "root.sg.test_0.d_0";
